change directory layout and add CMakeLists.txt
