+++ dbms/CMakeLists.txt
@@ -310,6 +310,9 @@ if (ENABLE_TESTS)
     if (ENABLE_TIFLASH_DTWORKLOAD)
         target_link_libraries(bench_dbms dt-workload-lib)
     endif ()
+    if (ENABLE_TIFLASH_STRESSWORKLOAD)
+        target_link_libraries(bench_dbms stress-workload-lib)
+    endif ()
 
     add_check(bench_dbms)
 endif ()
