+++ dbms/src/Storages/Page/stress/workload/PSBackground.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include <Common/MemoryTracker.h>
-#include <PSBackground.h>
+#include <Storages/Page/stress/workload/PSBackground.h>
 #include <Poco/Logger.h>
 #include <Poco/Timer.h>
 #include <fmt/format.h>
