+++ dbms/src/Storages/Page/stress/workload/PSWorkload.cpp
@@ -14,7 +14,7 @@
 
 #include <Common/MemoryTracker.h>
 #include <Encryption/MockKeyManager.h>
-#include <PSWorkload.h>
+#include <Storages/Page/stress/workload/PSWorkload.h>
 #include <Poco/Logger.h>
 #include <Storages/Page/V2/PageStorage.h>
 #include <Storages/Page/V3/PageStorageImpl.h>
