+++ dbms/src/Storages/Page/stress/workload/ThousandsOfOffset.cpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <PSWorkload.h>
+#include <Storages/Page/stress/workload/PSWorkload.h>
 
 namespace DB::PS::tests
 {
-    class ThousandsOfOffset : public StressWorkload
+class ThousandsOfOffset : public StressWorkload
     , public StressWorkloadFunc<ThousandsOfOffset>
 {
 public:
